refactor(chat): extract tool display subsystem from ChatBubble.kt
ChatBubble.kt was a 2578-line god file. Split it into four files in the
same package (zero import churn, no behavior change):

- ChatBubble.kt: user/assistant/system bubbles + dispatch + shared
  helpers (formatTimestamp, buildHighlightedString, InlineAttachment)
- ParsedToolData.kt: the parsed tool payload model
- ToolResultParser.kt: parseToolOutput + per-tool formatters + the
  Gson-compat JSON shims and tolerant exit_code parser
- ToolBubble.kt: ToolBubble + ExpandedToolContent + HeaderRow +
  SecurityRiskChip + CopyButton

formatTimestamp became internal (shared with ToolBubble.kt).
Verified: ktlintCheck + testDebugUnitTest green locally.
