+++ app/src/main/java/com/m57/hermescontrol/ui/chat/ParsedToolData.kt
@@ -1,0 +1,23 @@
+package com.m57.hermescontrol.ui.chat
+
+/**
+ * Cleanly parsed representation of a tool call, extracted from the
+ * tool.complete JSON payload (which contains both args and result).
+ */
+data class ParsedToolData(
+    val toolName: String = "",
+    val args: Map<String, Any?> = emptyMap(),
+    val result: Map<String, Any?> = emptyMap(),
+    val isTerminal: Boolean = false,
+    val stdout: String? = null,
+    val stderr: String? = null,
+    val exitCode: Int? = null,
+    val error: String? = null,
+    val summaryText: String? = null,
+    val durationSec: Double? = null,
+    val mainOutput: String? = null,
+    val extraFields: Map<String, String> = emptyMap(),
+    val isRunning: Boolean = false,
+    val diffOutput: String? = null,
+    val diffPath: String? = null,
+)
